requirements.txt
pandas
requests
beautifulsoup4
torch
torchvision
transformers
scikit-learn
Pillow
streamlit
plotly